requirements.txt
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.140.0
google-auth==2.32.0
google-generativeai==0.7.2
gspread==6.1.2
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
